move shared code to helpers
