+++ src/client/datascience/jupyter/kernels/kernelSelector.ts
@@ -20,7 +20,7 @@ import { IKernelFinder } from '../../kernel-launcher/types';
 import { reportAction } from '../../progress/decorator';
 import { ReportableAction } from '../../progress/types';
 import { IJupyterKernelSpec, IJupyterSessionManager, IKernelDependencyService } from '../../types';
-import { JupyterKernelSpec } from './jupyterKernelSpec';
+import { createDefaultKernelSpec } from './helpers';
 import { KernelSelectionProvider } from './kernelSelections';
 import { KernelService } from './kernelService';
 import { IKernelSpecQuickPickItem, LiveKernelModel } from './types';
@@ -431,20 +431,7 @@ export class KernelSelector {
 
     // When switching to an interpreter in raw kernel mode then just create a default kernelspec for that interpreter to use
     private async useInterpreterAndDefaultKernel(interpreter: PythonInterpreter): Promise<KernelSpecInterpreter> {
-        // IANHU: Shared code with kernel finder
-        const defaultSpecName = 'python_defaultSpec_';
-        const connectionFilePlaceholder = '{connection_file}';
-        const defaultSpec: Kernel.ISpecModel = {
-            name: defaultSpecName + Date.now().toString(),
-            language: 'python',
-            display_name: interpreter.displayName ? interpreter.displayName : 'Python 3',
-            metadata: {},
-            argv: ['python', '-m', 'ipykernel_launcher', '-f', connectionFilePlaceholder],
-            env: {},
-            resources: {}
-        };
-        const kernelSpec = new JupyterKernelSpec(defaultSpec);
-
+        const kernelSpec = createDefaultKernelSpec(interpreter);
         return { kernelSpec, interpreter };
     }
