+++ src/client/datascience/kernel-launcher/kernelFinder.ts
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 'use strict';
 
-import { Kernel } from '@jupyterlab/services';
 import { inject, injectable, named } from 'inversify';
 import * as path from 'path';
 import { CancellationToken, CancellationTokenSource } from 'vscode';
@@ -14,6 +13,7 @@ import { IExtensionContext, IInstaller, InstallerResponse, IPathUtils, Product, 
 import { IInterpreterLocatorService, IInterpreterService, KNOWN_PATH_SERVICE } from '../../interpreter/contracts';
 import { captureTelemetry } from '../../telemetry';
 import { Telemetry } from '../constants';
+import { createDefaultKernelSpec, defaultKernelSpecName } from '../jupyter/kernels/helpers';
 import { JupyterKernelSpec } from '../jupyter/kernels/jupyterKernelSpec';
 import { IJupyterKernelSpec } from '../types';
 import { getKernelInterpreter } from './helpers';
@@ -27,14 +27,6 @@ const macJupyterPath = path.join('Library', 'Jupyter', 'kernels');
 const baseKernelPath = path.join('share', 'jupyter', 'kernels');
 
 const cacheFile = 'kernelSpecPathCache.json';
-const defaultSpecName = 'python_defaultSpec_';
-
-// https://jupyter-client.readthedocs.io/en/stable/kernels.html
-const connectionFilePlaceholder = '{connection_file}';
-
-export function findIndexOfConnectionFile(kernelSpec: Readonly<IJupyterKernelSpec>): number {
-    return kernelSpec.argv.indexOf(connectionFilePlaceholder);
-}
 
 // This class searches for a kernel that matches the given kernel name.
 // First it searches on a global persistent state, then on the installed python interpreters,
@@ -74,7 +66,7 @@ export class KernelFinder implements IKernelFinder {
         this.cache = await this.readCache();
         let foundKernel: IJupyterKernelSpec | undefined;
 
-        if (kernelName && !kernelName.includes(defaultSpecName)) {
+        if (kernelName && !kernelName.includes(defaultKernelSpecName)) {
             let kernelSpec = await this.searchCache(kernelName);
 
             if (kernelSpec) {
@@ -319,18 +311,7 @@ export class KernelFinder implements IKernelFinder {
     private async getDefaultKernelSpec(resource: Resource): Promise<IJupyterKernelSpec> {
         const activeInterpreter = await this.interpreterService.getActiveInterpreter(resource);
 
-        // This creates a default kernel spec. When launched, 'python' argument will map to using the interpreter
-        // associated with the current resource for launching.
-        const defaultSpec: Kernel.ISpecModel = {
-            name: defaultSpecName + Date.now().toString(),
-            language: 'python',
-            display_name: activeInterpreter?.displayName ? activeInterpreter.displayName : 'Python 3',
-            metadata: {},
-            argv: ['python', '-m', 'ipykernel_launcher', '-f', connectionFilePlaceholder],
-            env: {},
-            resources: {}
-        };
-        return new JupyterKernelSpec(defaultSpec);
+        return createDefaultKernelSpec(activeInterpreter);
     }
 
     private async readCache(): Promise<string[]> {
